moved sqlitehelper into provider package
