+++ library/src/main/java/net/grandcentrix/tray/provider/TrayProviderHelper.java
@@ -18,7 +18,6 @@ package net.grandcentrix.tray.provider;
 
 import net.grandcentrix.tray.TrayPreferences;
 import net.grandcentrix.tray.core.TrayItem;
-import net.grandcentrix.tray.util.SqliteHelper;
 
 import android.content.ContentValues;
 import android.content.Context;
